Make double precision the release conserved-state storage
The release stored U = (rho, rho u, E) in float32, where the per-step
continuity increment dt*(-dF/ds) ~ 2.6e-17 kg m^-3 s^-1 falls below half a
ULP of rho ~ 9e-10 in 619 of 660 cells: the slow density adjustment could
not accumulate at all, and with no compensated summation that was permanent
stagnation. The controlling ratio is the stratified layer's mass-loading
timescale over the CFL step, tau/dt ~ 3.5e7, against 1/u = 1.7e7 for float
and 9.0e15 for double.

The second reason is architectural. An audit of the local SWMF tree shows
every shipped BATS-R-US build sets PRECISION = ${DOUBLEPREC} (all 23
share/build templates, GPU included), Config.pl defaults to double, and
State_VGB / Source_VC / the fluxes are default real, so U^{n+1} = U^n + dt*R
is a float64 accumulation with no compensated summation and no deviation
formulation. SWMF's answer to this failure mode is a policy, not a
mechanism. Our float32 state was the SWMF-inconsistent component; this
corrects the prototype to the precision SWMF production already uses rather
than importing a new requirement into SWMF.

Precision is now one knob. chromosphere::Real is the scalar storage type
(double; float only under the default-off diagnostic CHROMO_STATE_FLOAT32)
and Vec is arma::Col<Real>; every stored physical scalar is declared in
terms of them, including the mesh metrics, the Grid physical constants, CFL,
limiter_beta and sim_time. Coherence is the point: the earlier
CHROMO_STATE_FLOAT64 experiment promoted Vec alone while mixture_pack_ghost,
the Hancock predictor row, the RHS row, the timestep row, the packed-energy
round and every Grid constant still cast through float, so that build was an
accidental hybrid that looked promoted and was not. CHROMO_STATE_FLOAT64 is
retired and CMake hard-errors on it. run_control.hpp deliberately keeps
double rather than Real: it is intentionally Armadillo-free and holds no
stored state.

Validation. chromo_tests 11647 passed / 0 failed in BOTH builds with no
tolerance loosened anywhere; the 29 failures previously attributed to double
precision were the incoherent-knob artifact and do not occur.
test_release_production_table_domain_corners is no longer skipped in either
precision - its ULP walk now steps in the storage type and nudges against
the solver's own EOS acceptance instead of a test-side recomputation of the
endpoint energy. ctest 9/9 in both trees. release_validation.sh passes:
default path bit-identical, 23507804 faces, 0 Godunov fallbacks,
termination=end_time. The float32 build reproduces the archived pre-cutover
4000 s run byte-identically across .txt, .gamma_diag and .faceflux, so the
refactor is semantics-neutral.

Two long-standing open issues close with it, each with a matched float32
control so the attribution is causal:

- The residual velocity field now converges in time. It previously did not
  converge as dt -> 0 (refining the step made the level-to-level difference
  larger; every CFL pinned at L1 ~ 1.0-1.2 m/s), which was float32
  cancellation. In double the ladder converges at order ~1.9 in L1 and the
  floor drops to ~7e-3 m/s at CFL 0.50, while the float32 control still
  shows apparent order -1.2 to -1.3. Retired in mechanism only: the CFL
  acceptance matrix was not re-run, so CFL 0.50 stands on its original
  evidence, and no positive m/s noise floor is established.
- The hydrostatic fixed point is second-order convergent, which could not be
  measured before. t=0 max|F^rho| is 1.78e-13 at N=500 (vs 1.08e-12 in
  float32), ~1e8 above the double round-off floor so it has no
  representation component, and the 5-step ladder runs 8.11e-1 -> 3.32e-4
  m/s across ISO_NS 48..2000 at unbroken order ~2 over a 41x refinement.
  float32 tracks it only to ISO_NS=250 and reverses at 2000. A double Roe
  ladder gives the same orders, so it is the discretization, not the flux.

Two corrections against overclaiming, both from the earlier experiment's
hybrid build and both withdrawn:

- The 1600-1850 km flux spread is 13 %, not 2 %. Base/top goes 2.811 ->
  1.003 and the column mass budget closes (predicted/observed -214.9 ->
  0.994, 619/660 -> 0/660 frozen), but a broad monotone ~13 % decline
  remains and relaxes only slowly (14.3 % at 500 s to 13.0 % at 4000 s). It
  is not the first-face artifact; it is duration/resolution.
- Evaporation observables move ~30 %, not ~3 %, and downward: window-mean
  mass flux 3.74e-10 -> 2.49e-10 kg m^-2 s^-1, mean window velocity 5.85 ->
  4.74 m/s, top-cell velocity 13.18 -> 9.50 m/s. A lower chromosphere that
  cannot drain sustains a mass supply the column does not have, so float32
  over-predicted the flux. Any evaporation number from a float32 run is high
  by of order 30 %.

Unchanged and still open: the localized first-interior-face artifact is
precision independent (|f_diff|/f_total 16.30 % -> 16.10 %) and in double the
base face is f_diff-dominated, i.e. a real reconstruction/ghost defect rather
than noise. The N500 long-duration grid-convergence limitation is a
resolution question and is untouched.

Evidence: docs/studies/numerics/state_precision_release_cutover.md.

+++ run_control.hpp
@@ -67,7 +67,7 @@ struct StepCapSelection {
 /// is rejected rather than silently replaced by a dangerous cap.
 inline StepCapSelection select_step_cap(const char* step_cap_env,
                                         const char* end_time_env,
-                                        float time_mult) {
+                                        double time_mult) {
     StepCapSelection out;
     if (step_cap_env && *step_cap_env) {
         errno = 0;
@@ -96,9 +96,16 @@ inline StepCapSelection select_step_cap(const char* step_cap_env,
         out.source = StepCapSource::EndTime;
         return out;
     }
-    // Legacy path, bit-for-bit the historical expression, guarded against a
-    // float overflow / out-of-range cast for absurd time_mult values.
-    const float legacy = std::max(10000.0f, 10000.0f * time_mult);
+    // Legacy path, the historical expression, guarded against overflow / an
+    // out-of-range cast for absurd time_mult values. Deliberately `double` and
+    // NOT chromosphere::Real: this header is intentionally free of Armadillo and
+    // of the solver's storage-precision knob (see the file comment), and nothing
+    // here is stored state -- it turns one CLI scalar into an integer step cap.
+    // The historical expression was float; the result is cast to long long, so
+    // the cap is unchanged for every sane time_mult and only a pathological
+    // many-digit value could shift the integer by one.
+    // CHROMO_T_END / CHROMO_STEP_CAP bypass this path entirely.
+    const double legacy = std::max(10000.0, 10000.0 * time_mult);
     out.cap = (!std::isfinite(legacy) || legacy >= 1.0e12f)
         ? kUnboundedStepCap : static_cast<long long>(legacy);
     out.source = StepCapSource::LegacyTimeMult;
